+++ packages/sdk/src/types/order/bid/domain.ts
@@ -1,6 +1,7 @@
 import type { AssetType } from "@rarible/api-client"
 import type { BigNumberValue } from "@rarible/utils"
 import type { ContractAddress } from "@rarible/types"
+import type { CurrencyId } from "@rarible/api-client"
 import type {
 	PrepareOrderRequest,
 	PrepareOrderResponse,
@@ -17,14 +18,18 @@ export type GetConvertableValueFunction = {
 }
 
 export type GetConvertableValueRequest = {
-	assetType: AssetType
+	assetType?: AssetType
+	currencyId?: CurrencyId
 	price: BigNumberValue
 	amount: number
 	originFees: UnionPart[]
 }
-export type GetConvertableValueResult =
-  { type: "insufficient" | "convertable", currency: AssetType; value: BigNumberValue } | undefined
 
+export type GetConvertableValueResult = {
+	type: "insufficient" | "convertable"
+	currency: AssetType
+	value: BigNumberValue
+} | undefined
 
 export type IBid = (request: PrepareBidRequest) => Promise<PrepareBidResponse>
 export type IBidUpdate = (request: PrepareOrderUpdateRequest) => Promise<PrepareBidUpdateResponse>